ADDED: AtomicCounter class. It implements a fairly portable integer data container class with atomic increment and decrement operations. CHANGE: Removed locks from around GUID generators in ObjectMgr. Now those GUIDS will be generated using the AtomicCounter.

+++ src/arcemu-world/GMTicket.cpp
@@ -54,7 +54,7 @@ void WorldSession::HandleGMTicketCreateOpcode(WorldPacket & recv_data)
 	// Remove pending tickets
 	objmgr.RemoveGMTicketByPlayer(GetPlayer()->GetGUID());
 
-	ticket->guid = objmgr.GenerateTicketID();
+	ticket->guid = uint64( objmgr.GenerateTicketID() );
 	ticket->playerGuid = GetPlayer()->GetGUID();
 	ticket->map = map;
 	ticket->posX = x;
